Adds broken and working code, updates lesson plan
Lesson plan is updated and I've separated the broekn code out into two
bits, the first being the good looking broken code to show that it's
easier to navigate even if it's broken, and the second is the code that
is all squished and impossible to read with the human eye without
intense strain. Stlye has been added so everyone can see how it relates
to theirs.